pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discord-presence-mask"
version = "0.1.0"
description = "Privacy-aware Discord Rich Presence for Windows (masks sensitive window titles before publishing state)"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Your Name"}]
license = {text = "MIT"}
dependencies = [
  "pypresence>=4.3,<5",
  "psutil>=5.9,<6",
  "pygetwindow>=0.0.9,<0.1",
  "pywin32>=306; platform_system == 'Windows'",
  "python-dotenv>=1.0,<2"
]

[project.scripts]
discord-presence = "discord_presence:main"
